correcting mistakes, design

+++ src/com/epam/jwd/apotheca/controller/UserManagerService.java
@@ -5,7 +5,6 @@ import java.util.List;
 import javax.servlet.http.HttpSession;
 
 import com.epam.jwd.apotheca.dao.api.UserDAO;
-import com.epam.jwd.apotheca.dao.impl.RecipeDAOImpl;
 import com.epam.jwd.apotheca.dao.impl.UserDAOImpl;
 import com.epam.jwd.apotheca.model.User;
 
@@ -37,12 +36,16 @@ public class UserManagerService {
 	public boolean createUser(User user) {
 
 		boolean result = false;
+		if ( UserDAO.NAME_SUPER_DOC.equals( user.getName() ) ) {
+			return false;
+		}
 		User newUser = userDAO.save(user);
 		if (newUser != null) {
 			user = newUser;
 			result = true;
 		}
 		return result;
+		
 	}
 
 	public UserDAO getUserDAO() {
@@ -62,7 +65,10 @@ public class UserManagerService {
 	}
 
 	public List<User> getUsers() {
-		return userDAO.findAll();
+		
+		List<User> users = userDAO.findAll();
+		users.removeIf(u -> UserDAO.NAME_SUPER_DOC.equals(u.getName()));
+		return users;
 
 	}
 
@@ -71,11 +77,22 @@ public class UserManagerService {
 	}
 
 	public User getUser(String name) {
+		
+		if ( UserDAO.NAME_SUPER_DOC.equals( name ) ) {
+			return null;
+		} 
 		return ((UserDAOImpl) userDAO).getUser(name);
+		
 	}
 
 	public User getUser(Integer id) {
-		return userDAO.findById(id);
+		
+		User user = userDAO.findById(id);
+		if ( user != null && UserDAO.NAME_SUPER_DOC.equals( user.getName() ) ) {
+			user = null;
+		}
+		return user;
+		
 	}
 
 	public boolean canPrescribe(User user) {
@@ -90,15 +107,20 @@ public class UserManagerService {
 
 	}
 
-	public boolean isRoleEnabled(User user, Integer roleId) {
+	public boolean isRoleEnabled(User user, Integer permission) {
 
-		return (user != null) && ((roleId & user.getRole().getPermission()) == roleId);
+		return (user != null) && ((permission & user.getRole().getPermission()) == permission);
 
 	}
 	
 	public boolean deleteUser(Integer userId) {
-
-		return userDAO.delete(userId);
+		
+		boolean result = false; 
+		User user = userDAO.findById(userId);
+		if ( user != null && ! UserDAO.NAME_SUPER_DOC.equals(user.getName()) ) {
+			result = userDAO.delete(userId);
+		}
+		return result;
 
 	}
